restructure visual tests
