+++ visual-tests/tests/features/.empty-test-template.ts
@@ -1,5 +1,5 @@
-import { test } from '../src/test-runner';
-import { helpers } from '../src/helpers';
+import { test } from '../../src/test-runner';
+import { helpers } from '../../src/helpers';
 
 test(__filename, async({ page }) => {
   const table = page.locator(helpers.selectors.mainTable);
