+++ visual-tests/tests/features/editors/dateEditor/manually-value-edit.spec.ts
@@ -1,6 +1,6 @@
-import { test } from '../../../src/test-runner';
-import { helpers } from '../../../src/helpers';
-import { selectCell, selectEditor, openEditor } from '../../../src/page-helpers';
+import { test } from '../../../../src/test-runner';
+import { helpers } from '../../../../src/helpers';
+import { selectCell, selectEditor, openEditor } from '../../../../src/page-helpers';
 
 /**
  * Checks whether it's possible to manually edit the cell value.
