+++ visual-tests/tests/features/editors/textEditor/redo-multiline-text.spec.ts
