+++ visual-tests/tests/features/editors/textEditor/redo.spec.ts
