+++ visual-tests/tests/features/editors/textEditor/undo-multiline-text.spec.ts
