+++ visual-tests/tests/features/expand-dropdown-menu.spec.ts
@@ -1,6 +1,6 @@
-import { test } from '../src/test-runner';
-import { helpers } from '../src/helpers';
-import { openHeaderDropdownMenu, selectFromDropdownMenu } from '../src/page-helpers';
+import { test } from '../../src/test-runner';
+import { helpers } from '../../src/helpers';
+import { openHeaderDropdownMenu, selectFromDropdownMenu } from '../../src/page-helpers';
 
 test(__filename, async({ tablePage }) => {
   await openHeaderDropdownMenu(2);
