+++ visual-tests/tests/features/filters/hovering-clears-the-focus-state.spec.ts
