+++ visual-tests/tests/features/filters/shift-tab-navigation-through-all-components.spec.ts
