+++ visual-tests/tests/features/filters/tab-navigation-from-submenu.spec.ts
