+++ visual-tests/tests/features/filters/tab-navigation-through-all-components.spec.ts
