+++ visual-tests/tests/features/render-in-css-transform.spec.ts
@@ -1,5 +1,5 @@
-import { test } from '../src/test-runner';
-import { helpers } from '../src/helpers';
+import { test } from '../../src/test-runner';
+import { helpers } from '../../src/helpers';
 
 test(__filename, async({ tablePage }) => {
   await tablePage.evaluate('document.body.style = "transform: scale(0.75);"');
