+++ visual-tests/tests/features/select-few-cells-by-mouse.spec.ts
@@ -1,6 +1,6 @@
-import { test } from '../src/test-runner';
-import { helpers } from '../src/helpers';
-import { selectCell, makeSelectionFromCell } from '../src/page-helpers';
+import { test } from '../../src/test-runner';
+import { helpers } from '../../src/helpers';
+import { selectCell, makeSelectionFromCell } from '../../src/page-helpers';
 
 test(__filename, async({ tablePage }) => {
   const cell = await selectCell(1, 0);
