+++ visual-tests/tests/features/tab-navigation.spec.ts
@@ -1,5 +1,5 @@
-import { test } from '../src/test-runner';
-import { helpers } from '../src/helpers';
+import { test } from '../../src/test-runner';
+import { helpers } from '../../src/helpers';
 
 test(__filename, async({ tablePage }) => {
   await tablePage.locator('html').press('Tab');
